move some files to chat folder
